+++ simplepyml/util/functions/loss/mse.py
@@ -1,5 +1,5 @@
 import numpy as np
-from .loss import BaseLoss
+from simplepyml.util.functions.loss.loss import BaseLoss
 
 
 class MSE(BaseLoss):
@@ -11,8 +11,8 @@ class MSE(BaseLoss):
         values: int | float | np.integer | np.floating | list | np.ndarray,
         expected: int | float | np.integer | np.floating | list | np.ndarray,
     ) -> np.ndarray | np.float64:
-        values = np.asarray(values)
-        expected = np.asarray(expected)
+        values = np.asarray(values, dtype=np.float64)
+        expected = np.asarray(expected, dtype=np.float64)
 
         return np.mean((expected - values) ** 2, dtype=np.float64)
 
@@ -21,6 +21,6 @@ class MSE(BaseLoss):
         values: np.ndarray | list,
         expected: np.ndarray | list,
     ) -> np.ndarray:
-        values = np.asarray(values)
-        expected = np.asarray(expected)
+        values = np.asarray(values, dtype=np.float64)
+        expected = np.asarray(expected, dtype=np.float64)
         return 2.0 / values.size * (values - expected)
